ajout de la validation de défis sur la page admin

+++ src/controleur/Facade.java
@@ -10,6 +10,7 @@ import javax.persistence.TypedQuery;
 import exceptions.ExceptionUserNonDefini;
 import modele.Defi;
 import modele.Defi_A_Valider;
+import modele.Defi_Valide;
 import modele.Demande_A_Rejoindre;
 import modele.FilActu;
 import modele.Groupe;
@@ -84,7 +85,18 @@ public class Facade {
 		return defi;
 	}
 
-	public void validerDefi() {
+	public void validerDefi(int id_dav) {
+		Defi_A_Valider dav = em.find(Defi_A_Valider.class, id_dav);
+		
+		Defi_Valide dv = new Defi_Valide();
+		em.persist(dv);
+		dv.setDefi(dav.getDefi());
+		dv.setGroupe(dav.getGroupe());
+		dv.setMembre(dav.getMembre());
+		
+		em.remove(dav);
+		
+		
 		
 	}
 
@@ -98,7 +110,7 @@ public class Facade {
 
 	public boolean ajouterMembre(String email, Groupe grp) {
 		TypedQuery<Membre> req = em.createQuery("select m from Membre m WHERE email='" + email + "'", Membre.class);
-		if (req.getResultList() == null) {
+		if (req.getResultList() == null || req.getResultList().size() == 0) {
 			System.out.println("Aucun résultat pour cette requête.");
 			return false;
 		}
@@ -106,7 +118,7 @@ public class Facade {
 		mb.getGroupesAppartenus().add(grp);
 		
 		TypedQuery<Demande_A_Rejoindre> req2 = em.createQuery("select d from Demande_A_Rejoindre d WHERE GROUPE_ID=" + grp.getId() + " AND MEMBRE_ID=" + mb.getId(), Demande_A_Rejoindre.class);
-		if(req2.getResultList() != null) {
+		if(req2.getResultList()!= null && req2.getResultList().size() != 0) {
 			Demande_A_Rejoindre dar = req2.getSingleResult();
 			em.remove(em.find(Demande_A_Rejoindre.class, dar.getId()));	
 		}
@@ -286,7 +298,7 @@ public class Facade {
 		System.out.println("Groupe : " + g);
 		System.out.println("Membres : " + g.getMembres());
 		
-		g.getMembres().add(mb2);
+		//g.getMembres().add(mb2);
 		
 		System.out.println("Membre.getGroupes " + mb2.getGroupesAppartenus());
 		System.out.println("Admin.getGroupesAdmin " + mb.getGroupesAdministres());
@@ -311,6 +323,7 @@ public class Facade {
 		dav.setGroupe(g);
 		em.persist(dav);
 		
+
 		Membre mb3 = new Membre();
 		mb3.setCoeff_sportif(1);
 		mb3.setEmail("cha@sfr.fr");
@@ -319,7 +332,12 @@ public class Facade {
 		mb3.setPrenom("Charlotte");
 		em.persist(mb3);
 		
-
+		Defi_A_Valider dav2 = new Defi_A_Valider();
+		dav2.setDefi(d);
+		dav2.setMembre(mb3);
+		dav2.setGroupe(g);
+		em.persist(dav2);
+		
 				
 		Demande_A_Rejoindre dar = new Demande_A_Rejoindre();
 		dar.setGroupe(g);
